fix(marketing): keep render state route static
